Completed problem #800
I've elected to release the code I used for solving problems outside of
the first 100. I've seen other repos do it, and I don't outright provide
the answer, so it should *probably* be fine. If Project Euler contacts
me to request it be taken down or something similar, I am completely
fine with removing it.